// packages/router/src/models.d.ts
export type Routes = Route[];
// ...
export interface LoadedRouterConfig {
    routes: Route[];
    injector: EnvironmentInjector | undefined;
}